Formatting, add scopes, add next and back buttons
