+++ web/client/src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 import MaterialUI from "./components/layouts/Material";
-import { TokenFormat } from "./components"
+import { TokenForm } from "./components";
 // import "./App.css";
 // import Main from "./components/Main";
 // import { Route, HashRouter } from "react-router-dom";
@@ -50,7 +50,7 @@ function App() {
     //   <Route path="/Scopes" component={Scopes} />
     // </HashRouter>
     <MaterialUI>
-      <TokenFormat />
+      <TokenForm />
     </MaterialUI>
   );
 }
